Initial migration of the code from transition-legacy
This code was in a private repository, where other applications were
developped as separate branches or directories. This commit contains the
part of that other repository to open source with the Transition
application.

The code from the legacy repo was contributed by the following
contributors, ordered by total number of commits.

Co-authored-by: Pierre-Léo Bourbonnais <leo.bourbonnais@polymtl.ca>
Co-authored-by: Geneviève Bastien <gbastien@versatic.net>
Co-authored-by: Alexis Viallard <alexis.viallard@polymtl.ca>
Co-authored-by: Julien Faucher <julien.faucher@gmail.com>
Co-authored-by: Vincent Chabin <vincent.chabin@polymtl.ca>
Co-authored-by: khadidjaSeridi <seridi_k@yahoo.com>
Co-authored-by: Yannick Brosseau <yannick.brosseau@gmail.com>
Co-authored-by: Alexis Martin <alexis.martin.2@ens.etsmtl.ca>
Co-authored-by: Yasmine Moumou <moumouyasmine@gmail.com>
Co-authored-by: Ian Letourneau <letourneau.ian@gmail.com>