json2capnp: Make the cache path argument mandatory
Now that the project shortname is not an environment variable anymore
and that the project directory can be anywhere and not only in a
projects/<projectShortname> directory, it is not possible de know this
path from the .env file data. The absolute path must be set on the
command as second argument

For example, to run json2capnp for the example app, where the cache is
in `examples/runtime/demo_transition` directory, the service can be run
as `yarn start:json2capnp -- 2000
$PWD/examples/runtime/cache/demo_transition`, or `cd services/json2capnp
&& cargo run -- 2000 ../../examples/runtime/cache/demo_transition`.

The dotenv and home packages are not required anymore.