setup: Drop operator family as well as class
Fixes #457

In recent postgresql versions (at least on 14+), dropping the operator
class is not sufficient to clean up everything and when re-creating it,
there's an exception for a duplicate key in the pg_amop table. This
prevents the setup task to be run more than once.

Dropping the operator family fixes the issue. It works also on old
versions of postgresql.