config: Add default paths for project config file
Fixes #425

First, we check if the PROJECT_CONFIG environment variable contains the
path to the config file. Otherwise, it is expected to be in the
$homedir/.config/transition/config.js or /etc/transition/config.js.

The path in the .env now has to be absolute or relative to the initial
current working directory where the command was run. The ${rootDir} does
not apply anymore.

Add a start:demo script in the package.json which sets the
PROJECT_CONFIG environment variable to the example config.