Use node's inspector feature instead of inspector api
This removes a dependency, which was triggering `yarn` warnings anyway.
They both produce the same kind of output.

It was tested with node 16 and 12.9 (our minimal version) and the
node inspector feature is available in both, at least on linux.

Also append the date to the profiling data file and add it to the
.gitignore.