Reduce the amount of logging in batch calculations
Each odTrip calculated was logged when read and when calculated. For
large input files, that is a lot of log lines, that are useless for
debugging and any actual error logged during the execution is kicked out
of the log buffer.

Log entries at read time are shown only if the debug option is set to
true. Execution logs are shown for every 1% progress.

Same 1% logging is added for the accessibility map batch calculations.