Upgrade versions of prettier and eslint
Some newer syntax of typescript were not readable by the current formatter
(like `export * as something...`). All eslint and prettier packages are
upgraded to their latest version.

`yarn format` is run on the code, some of the rules have changed and the
files formatted.

Also remove the `eslint --fix .` from the `format` script, as it is
included in prettier-eslint.