Fix the Collapsible lazy rendered elements not displaying
Fixes #494

This bug occurred since the upgrade to react-collapsible 2.10

On first click to a lazy rendered element, the height is 0 and the
children are not expanded. Only on third click did it show. It is a bug
in the react-collapsible library (see
https://github.com/glennflanagan/react-collapsible/issues/36)

To bypass this, add a 1px padding on top of the collapsible's inner
content, to make sure the chilren have a height.