trRouting: Add the backend routes for the `route` query
The `route` method of the trRouting backend service is renamed to
`v1TransitCall` and a new `route` method is added to receive the
TransitRouteQueryOptions object and return a route response.

The `trRouting/route` express route and the `service.trRouting.route`
socket route are added to the backend to call the trRouting's backend
service `route` method.