trRouting: Add a new `route` method to TrRoutingService
The old `route` method is renamed to `routeV1` and all callers use it
instead. This method still supports all of route, accessibility maps and
od trips queries.

A new `route` method is added, that only calls the `route` functionality
of TrRouting, with a thinner API. Tests are added for this new methods.
Mocks and stubs are updated accordingly.