trRouting: Let the calculator use the new `route` method
The return type is changed to a thinner object, but which needs to be
properly converted before showing any result to the user.

The batch routing use the result converter to convert the result to a
"human-readable" object to export to csv.

The routing result UI is updated to format the times from seconds to
human-readable times.