Upgrade knex dependency to 2.4.2
The main difference is in how to import the knex objects from 'knex', as
named imports instead of the default one.

Also, return types from queries, when only one field is selected is now an
object instead of the value directly. The queries have been updated
accordingly.

The sequential tests all pass. They had been broken a while back, since
the line's agency ID is mandatory. They are fixed in this commit.