batch routing: Split the calculation from attributes
The TransitBatchRouting object has 2 responsibilities: calculation and
attributes management.

The calculation has been moved to a TransitBatchRoutingCalculator class
with a static method, which takes a demand and query parameters as arguments.

The batch attributes management actually represents the configuration of
a demand file, in csv format, so it was moved to a transitDemand
service, in a class called TransitDemandFromCsv, which better represents
what is really is.

The frontend has been updated to use the new classes.