batchCalculation: Add panel and dashboard contribution
The batch calculation is part of a new supply/demand interaction
analysis module, so a new dashboard contribution is added for it.

It adds the main panel, which displays a list of 'batchRoute' tasks, the
same as for the batch routing functionnality. There's also a button to
create a new analysis.

The new calculation form skeleton is added, but contains only a back
button.