FileImportForm: Use the FileUploaderHOC in this component
Instead of each individual import form that use the FileImportForm
passing the HOC's props, the individual import forms do not need the
HOC's props and can just contain a FileImportForm. This last components
uses the HOC.

Also, the FileUploaderHOC can register itself to the proper events
instead of passing props and relying on the consumers to call these
methods.