fileUploaderHOC: Add a prop to track status of upload
Add an uploadStatus props, which passes the state of the FileUploaderHOC
with the status of the file upload: notUploaded, uploading (with a
progress), error, completed or aborted.

This prop can be used by consumer to track the upload status and know
when the upload is completed.

Also, do not pass the chunkSize in the socketIOFileClient constructor, as it
is not one of the possible options and it does not have any effect, the
chunk size comes from the server.