UI: Display split lines passing by a same segment
Offset transit paths on the map to better see where each goes. Offset
stop nodes to be in the middle of crossing paths to see which paths have
a stop there.

It is inspired by this article for by the transit app developers:
https://blog.transitapp.com/how-we-built-the-worlds-prettiest-auto-generated-transit-maps-12d0c6fa502f/

But instead of image processing, the geojson data is processed directly
and modified before display.

This feature is experimental and needs the `features.map.prettyDisplay`
preference to be set.

Also, the transit nodes by which paths pass are placed such that they
are in the middle of the lines.

The overlapping segments take time to calculate, it is done
asynchronously. After zooming or panning, it may take a few seconds
before the map shows the pretty path

TODO:

* Improve performance of the overlapping lines
* Cache the overlapping segments to avoid having to recalculate on every
  map refresh
* Find a better sorting algorithm to decide the order of he ofsetted
  lines
* Add a use configurable minimum zoom level for which the algorithm is
  applied (now hardcoded zoom limit of 14)
* There are still a few bugs in the display
* Add more tests
* Make clearer which lines stop at nodes or not (for example a segment with
  express and non-express routes)

Limits:

* The current offset is applied on a perfect match between the overlapped segment coordinates and the line coordinates. This can be adjusted in turf.lineOverlap function.

This project was done as part of a final project by a team of computer
engineering students at Polytechnique Montreal, in the winter 2023
semester. Thank you for the contribution!

Co-authored-by: Wassim27 <mw-guellati@outlook.fr>
Co-authored-by: Gabriel Bruno <gabriel.bruno@polymtl.ca>
Co-authored-by: Florence Lauer <florence_lauer@outlook.com>
Co-authored-by: MohamedAli-M <mohamed-ali.messedi@polymtl.ca>
Co-authored-by: mahdiguermache <mahdi.guermache@polymtl.ca>
Co-authored-by: nik498 <54679682+nik498@users.noreply.github.com>