Wait for cache creation before the socket server starts
Fixes #597

Batch calculations require the cache to exist. If we run the cache
creation asynchronously, tasks may be restarted/enqueued before it is
completed and the results will be wrong at best and potentially
invalid if the trRouting server starts with incomplete data. We should
wait for the cache to be completely saved before allowing client
connections.