jobs: Create the job with its input files in the resources
Currently, the files were only written once the job is completed, with
outputs and inputs file. But the input file is known from the beginning
and should be part of the job's resources from the start. This also
ensures that the file is known if a running job is duplicated.