jobs: Support renaming input files when creating jobs
Input files can come from uploads by the user. To avoid having to manage
the user's uploads, the files are often named a unique name and the next
upload replaces it. But in the jobs, they can be renamed to their
original names, so the user can track which files was uploaded exactly.