dataSources: Support collection and single queries with owner
Fixes #586

The dataSource collection db query, as well as the read and findByName, now
receives the userId as extra argument.

The getDataSource backend method, which either creates or gets a
datasource also has a userId parameter, such that only the data sources
owner by the user or with no owner are returned.

Update callers and tests