events: Add a ClientEventManager class for events global to a user
In many places in the server code, we use the socket as event listener.
But the socket object does not survive and browser tab shutdown, user
disconnect/reconnect, etc. For long running server actions, like
executable jobs, the user want to continue to receive notifications
event after disconnecting.

This connection allows to register event emitters for a specific user.
Long running tasks can use the event emitter provided by this class to
notify events and automatically, all the registered EventEmitter objects
for the user will emit and event and the listeners will be notified.

This commit introduces the ClientEventManager class and global object
and lets the socket be registered/unregistered after connection or
disconnection.