jobs: Let the ExecutablJob emit events on the global emitter
The EventEmitter in argument to the create, save, enqueue and delete
methods is now optional, and if not set, retrieved from the global
clientEventManager for the job's user.

This is the first part for fixing #589. For now, the progress of the
jobs will be displayed in the notifications after reconnection, as if
the user was never disconnected.