config: Fetch config from server in frontend
Instead of having it built into the application at webpack time, it is
fetched at load time from the server.

This adds an express route named `/config` that returns a Status with
the project configuration.

The client's entry page now needs to asynchronously fetch it from server
using this route and the page is only rendered when the response is
back. If there is an error while getting the configuration, the
MaintenancePage is displayed.

Because the initial frontend configuration is now asynchronous at the
application start, the i18n.config is converted to a function, which
initializes the i18n object that should be called once the configuration
is fetched.