batchCalculation: Add a table and queries to save the batch route results
This table saves the raw results from the calculation. There is one row
per job/calculation. For now, the result is simply a json object,
serialization of the ResultByMode object, with keys being the mode
calculated and the value is the parameter object used to build either
the TransitRoutintResult or UnimodalRouteCalculationResult.