batchRoute: Save results to DB and generate files at the end
The routeOdTrip function simply returns the ResultByMode or error object
for the odTrip to calculate.

The batchRoute function stores this result to the database.

The functions to generate the CSV, detailed CSV and geojson files are
moved from the TrRoutingOdTrip file to TrRoutingBatchResult, which
handles all result management.

After all calculations are completed, the files are generated from the
results in the database, after which the results are deleted.

Generating then deleting results is done at the end of the job, and not
when the user asks for it, to make sure the data from which results are
calculated is still valid. Especially for the geometries, generating the
result files long after the calculation has ended increases the risk
that the transit objects used for calculation have changed.