config: Provide a function to load config in chaire-lib-backend
To allow individual applications like Transition and Evolution to each
have their own configuration options and register them at load time, the
configuration is not loaded directly in chaire-lib-backend, but requires
an `initializeConfig` function to be called before hand.

A new file, called `loadConfig.ts` in the config directory of
chaire-lib-backend allows to directly call the configuration
initialization, as it was done previously in the config file. This file
can be used by tasks and migration knexfiles, that will run, but won't
be run in the context of the server.

A new configuration file named `app.config.ts` is added to the
transition-backend package. This one will be responsible to load the
Transition application's configuration. For now, it only calls the
chaire-lib's setup, but eventually will add its own configurations.

Since the configuration files also make sure to initialize the dotenv
environment variables, calls to .env in other files are now unnecessary.