i18n: Remove initial language check
This language check caused the defaultLocale to always be set when
reloading the page instead of using the detection mechanism. It
previously worked because the config used in the i18n file was from
chaire-lib-common, which was imported before getting the actual config
from chaire-lib-frontend. When correctly importing the frontend config,
or if setting default values for the languages in the common config,
the issue arose.