chaire-lib: Move the FrontendUser type to common workspace
Fixes #630

The type is renamed to CliUser.

The type itself is not frontend-specific, it contains only the
non-private user information, as well as the required data to verify
permissions as required. It can be used by any client-side code and not
only in the react frontend.

Its current implementation remains in chaire-lib-frontend though (the
toCliUser function in `chaire-lib-frontend/lib/services/auth/user`).