tests: Mock the db.config file
Fixes #633

In unit tests testing files that import database query files, the
database configuration is always loaded and requires the .env file to
have been parsed with proper data. This should not be a requirement to
run tests, especially that the database queries that are under test are
already mocked. This adds a mock for the db.config, which simply
initializes a dummy knex object for 'sqlite' that should make the tests
pass without configuration.

The `db.config` file is automatically mocked in chaire-lib-backend and
transition-backend packages.