auth: Add an interface for authentication model
Add a IUserModel and IAuthModel interfaces, which should be implemented
by any authentication model. These interfaces allow various user model
and authentication model implementations, with the same basic methods.
There is a AuthModelBase and UserModelBase implementations with the
default common user attributes.

The current UserModel's static methods are moved to the UserAuthModel
class, which is the default implementation, where the applications
users reside in the database's `users` table. And the UserModel now
extends UserModelBase. A few of its method are now part of the base
interface.

Update tests to work with this new approach.