dependencies: Update i18next package versions
* i18next: from 20.2.2 or 19.5.1 to 22.4.15
* i18next-fs-backend: from 1.1.1 to 2.1.1
* i18next-browser-languagedetector: from 5.0.0 to 7.0.1
* i18next-http-backend: from 1.0.16 to 2.1.0
* react-i18next: from 11.7.0 to 12.2.2

With the i18next major release to 22.0, typescript types have been much
updated. The main change is that TFunction can now optionally return
null, and it allows null by default. Since most of our usages of t
functions are expected to be `string | undefined`, we can set the
`returnNull` option to `false`, but for compilation to work correctly,
it needs to be typed. As suggested in the i18next documentation, we add
a i18next.d.ts type file with the returnNull set to false.

The setDerivedStateFromProps static methods on translatable components
do not compile correctly anymore. Since it should not be used anyway
(see issue #307), the prop type is just set to `any` for now.