preferences: Support loading preferences from socket or fetch
Add a `/load_user_preferences` express route to retrieve the preferences
for the current user. This route, as well as the `preferences.read`
socket route return a status with the preferences as object.

Add unit tests for this purpose.

The fetch route supports any authentication model, not just the `User`
model.