preferences: Support listeners to changes
Fixes #666

Instead of sending an event emitter to update and load calls, typically
coming from the serviceLocator, now consumers have the possibility to
register to the Preferences class to be warned of changes with the
`addChangeListener` and `removeChangeListener` methods.

These methods are called after loading and updating preferences, with
the newly loaded or updated preferences as parameters.

Add unit tests for these 2 methods.