zones: add DB table, queries and classes
First part of #689

A zone is an area (polygon or multipolygon) on the map. It has a
geography, as well as name, shortname, id and data.

Zones are associated with a data source of type 'zones'.

This adds the migration to add the tr_zones table, as well as queries to
create, read, update, delete and fetch collection of zones.

It also adds a Zone and ZoneCollection classes.