mainMap: Emit 'map.loaded' event on map error
When working offline, without internet access, the mapbox style cannot
be fetched from mapbox and the main map will be in error. The 'load'
event won't be called and the 'map.loaded' event, which triggers the
data fetch is never called, either, causing the application to be empty,
but the user is still able to go to the various menus.

We add a listener on the 'error' event of the map and if the map is not
loaded already call the 'map.loaded' event. Even if the map data cannot
be displayed, the data can still be navigated in the application. It
should not display an empty application.