auth: check is username exists and use part of a uuid as name
fixes #711

Instead of 6 digits, we use the last 10 characters of a uuid (hex
characters) for the "random" part of the username. It also checks if the
username exists before assigning it, up to 10 tentatives, after which it
fails with a message that a unique name cannot be found.

According to https://zelark.github.io/nano-id-cc/, with the hex alphabet
and 10 characters, it should take 148K IDs to have a 1% probability of
collision. With 10 checks, it should be enough for most applications
using chaire-lib.