Small UI fixes (#751)
* Use fixed-size elements to display the coordinates in the top toolbar.

Otherwise, we get annoying flickers when some coordinates have fewer decimals. Also, set align-items to 'center' so that the coordinates are aligned with the other buttons.

* Add missing copyright header

* Change react-collapsible's arrow angles

This now follows the more common convention of "right" being "closed" and "down" being "open".

* Fix "lignes" header having two background colors

* Use fixed-size elements to display the coordinates in the top toolbar.

Otherwise, we get annoying flickers when some coordinates have fewer decimals. Also, set align-items to 'center' so that the coordinates are aligned with the other buttons.

* Improve <select> styling

Add rounded corners, a thinner border and set the same font size as <input> components

* Add a custom white chevron for the select component

Otherwise, it's black on black and we can't see it.