deck.gl: Support events on map and map layers
* Change the API of the event handlers
* Drop support of mapbox types in event handling
* Change the events to use the new API
* mouseDown/Move/Up are now onDrag and onDragEnd events, which happen
  only with a specific feature, these are easier
* There are simple tooltip events, which simplify popup management when
  only a simple text is required
* The `click` event is separate in `leftClick` and `rightClick`

TODO: The `path.hoverNode` and `path.unhoverNode` are not yet
functional, as they programatically need to create a popup or menu at a
specific location.

TODO2: Aesthetic changes are not taken into account, as zoom events (and
other layout events) are not yet properly handled

TODO3: Not all layers are refreshed, even though the collection they
display has elements that have changed. For example, saving a path does
not show the updated path. Deck.GL does a shallow comparison to
determine if the layer needs to be refreshed and probably the shallow
comparison does not show any change if the coordinates changed.