eventManager: Add methods to listen and emit typed events
This allows to use a type for the events, with a `name` and `args`
fields for the name of the event and types of arguments in the callback.

The methods are generic, with the generic type being the type of the
event. The listening method is `onEvent`, the emitting one is
`emitEvent`. It can be called in the code with

`eventManager.emitEvent<MyEventType>('event.myevent', myArgs)`
`eventManager.onEvent<MyEventType>('event.myevent', (myArgs) => void)`

Add unit tests for these new methods.