dashboard: Add a ref object to the right panel div
Add a `PanelSectionProps` type, which extends the `LayoutSectionProps`
but adds an optional ref to its parent.

The right panel now has a ref on his `div` element and passes it down to
its children, so they can keep the scroll state of the element.

The TransitionDashboard contributions receive the parent ref and can
pass it to their children.