transitNodes: All a button to delete all unused transit nodes
fixes #808

Add a `transitNodesUtils` file with a function to delete unused nodes,
either an array of node ids or all nodes and reload the collection.

Add a button in the `TransitNodePanel` to delete all unused nodes, with
a confirmation modal before doing so.

Also change the current delete on selection to use the node util's
function and delete all selected nodes at once.