osrm: Use the host even for local servers
fixes #821

This uses the value set in the mode config's host field even for local
servers. This allows to specify a specific URL on which to join the
server, for example an ipv4 url where 'localhost' would have resolved to
an ipv6 url, or simply a named URL for the local server.