gtfs: Add the possibility to infer frequency based schedules at import
fixes #718

If the option is selected, instead of importing exact trips from the
GTFS, it will try to infer a frequency based schedule from the trips in
the period.

For each direction, it calculates the possible frequency as `(time of the
last trip - time of the first trip) / (number of trips - 1)` or 60
seconds if there is only one trip.

Custom start and end time will be used respectively if the time between first
trip and last trip and the period boundary is superior than the frequency.
Otherwise, it uses the period's official boundary for start or end.

If more than one path is used by the schedule for a direction, the most
frequent one will be used for the whole schedule.