trRouting: Use threaded version
Add mecanism to add --threads option when starting trRouting

In default normal mode, we start with 2 threads.

We add a new startBatch and stopBatch functions which will replace the start/stopMultiple. Those
will start just one trRouting with the specified number of threads instead of starting multiple processes

We added a check in the startBatch function to insure we do not start more thread than the config limit

Added a few unit test