gtfs: Always use custom times when inferring frequencies
This seems to be currently more intuitive for end users, instead of
using the period start/end if the diff between the time and closest trip
time is lower than frequency.

Given various issues with schedules (#681, #854, the fact that all trips
start at the exact period start, problems inferring frequency with
asymetrical trips, to name a few), this sounds like an
acceptable compromise.