nodes: Add a backend method to save individual node
and update its transferable nodes.

When a node's geography is updated, the transferable nodes FROM the node
will be calculated. Also, all previous nodes that transferred to this
node will be updated, as well as all nodes now accessible. Only the
values for the current nodes are updated, all other transferable nodes
are left untouched.