chaire-lib: A new object in editing mode should be considered new
fixes #857

Even if there is no history, a new object that is being edited should be
considered new. If the default values are invalid, the `validate` method
should return `false`. This will fix the saving of the new nodes saved
with default values.

Update the ObjectWithHistory test to test all possible cases of
new/updated editing/not editing objects.