db: Allow create/update/delete queries to receive a transaction
For default create/update functions, the last argument (the `returning`) is
changed to an option parameter, which can also receive a transaction
object, in which to include the current query.

The default delete functions receive an `options` argument with an
optional transaction.

The current concrete query functions do not yet use the transaction
parameter.