Don't recalculate transferable nodes on startup
We now save the transferable nodes in the database, so we don't really need to recalculate them at
each startup. But we still need to save them to the capnp cache file.

We add a saveAllNodesToCache function to NodeCollectionUtils, based on saveAndUpdateAllNodes, that
only store the node data without recalculating it.
We add a else in the recreateCache function where we test for the refreshTransferrableNodes flag to
call our new function in that case. (That way we always save the transferable nodes to cache)

The dbToCache were missing a mockClear, we switch it to jest.clearAllMocks() and fixes the erroneous tests

The call in socketServerApp was changed to ask for not refreshing the transferable node when recreating the
cache on startup