GTFS: Use postgis to get existing nodes when aggregation radius is 0
fixes #908

Since KDBush sometimes does not return a node which is at the exact
location as another one, the main code path sometimes created new nodes
at the exact same location when the aggregation radius was set to 0.
postgis, on the other hand, does not do that, so we get the existing
node from the database in that case.