db: Let transitScenarios queries support transaction
All individual queries are also transactions, for the scenario and its
associated services. Those queries can receive an optional transaction
option parameter, that if set, will be used instead of creating a new
transaction.

The queries are also modified to eliminate the read from DB/process
locally/write to DB flows. Transactions should be made as short and fast
as possible, so single write SQL queries should be favored.

Add a unique index on the scenario/service pair in the
`tr_scenario_services` table.