routing: Move the `TransitRoutingResult.ts` file to chaire-lib-common
The file is moved and references updated. Since it depended on a
pathCollection object to get the geojson and chaire-lib does not have
access to this data, the getPathGeojson function now receives a function
that translates a transit step to geojson instead of the PathCollection

In transition-common the `TransitRoutingResult` now provide a class that
takes a PathCollection in the constructor and can be used as the
segmentToGeojson function. Transition calculations can thus use this
class to retrieve geojson path. Other tools can provide their own
function, or do not require the geojson result.

This commit is part of work to move the routing calculations to
chaire-lib instead of Transition, so that other tools can benefit from
it.