routingResult: Drop the concept of walkOnly route
The `getWalkOnlyRoute` function is not part of the `RoutingResult`
interface and needed separate code paths to deal with it. Instead the
`TransitRoutingResult` simply return the walk only path in the `getPath`
function at the proper index.

Callers can then use the `pathIsRoute` typeguard to differenciate if
necessary.

This commit is part of work to move the routing calculations to
chaire-lib instead of Transition, so that other tools can benefit from
it.