routing: Add routing query types and validation function
2 types are specific for transit queries: `TransitRoutingBaseAttributes`
and `TransitRoutingQueryAttributes`. They are used to describe the
extra parameters for transit calculations.

The `RoutingQueryAttributes` type describe any calculation attributes,
while the `TripRoutingQueryAttributes` type describes a single
calculation for a single origin/destination. The `withAlternatives`
attribute has been moved to the `RoutingQueryAttributes` as any routing
can provide alternatives, it is not specific to transit.

These types are used for Routing calculation queries and is used to pass
query data between client and server. So there's a helper function
that exhaustively validates the routing query attributes before creating an
object which has all required fields.

Those types are meant to replace the `TransitRoutingAttributes` type,
at least in the calculations in chaire-lib and in the backend. It is
less Transition oriented and can be used in other calculation contexts,
like surveys.

This PR updates