map: Add a preference for the opacity of xyz tile layer
fixes #966

This moves all map related preferences to a separate section in the
preferences edit form.

Add a preference called `mapTileLayerOpacity`, which is a value between
0 and 1 representing the percentage opacity of the xyz tile layer.

Extract the tile layer getter in the `TransitionMainMap` so it can be
reset when the layer opacity is changed. If the opacity is 0
(invisible), the layer is not present.