routing: Support alternatives for single-mode routing calculations
Add a `RouteParameters` type to replace the `MapMatchParameters` in the
`route` function, as those 2 parameters types are not expected to be the
same and this clarifies the difference. Add a `withAlternatives`
property to this new type.

The `TripRoutingQueryAttributes` object is added as a parameter to the
`getRouteByMode` function call, which allows to set the
`withAlternatives` property of the `RouteParameters` object.