Autostart TrRouting on server startup
With the use case of the API, we might have user that will not login in the web interface
to manually start TrRouting.

Since we now recreate the cache automatically on startup, we can also start TrRouting.

We use a delayed await, so that we can do other stuff while trRouting start

We removed the line that would kill other TrRouting, since we now use the restart function