Add configuration flag to enable use of the --cacheAllConnectionSets flag of trRouting
We add an option entry in the config file "trRoutingCacheAllScenarios" which default to false.

If set to true, this will add the cacheAllConnectionSets to the call to trRouting to let it cache the
connections for all scenario. By default we only cache the last used. If many scenario are used, this will
use quite more memory

This only apply to the "normal" trRouting, not the batch one. Usually, batch calculation only use one scenario,
it it would be irrelevant.

We used to opportunity to add types to all functions in TrRoutingProcessManager