Move the await outside of the for loop in the Routing::calculate function
The await in the for loop made the call to each RoutingMode sequential. The the promise generation and
the promise.all after the loop, we can make all the API call in parallel. Those are mostly
fast OSRM calls which are quite fast, but we can still have a small gain.

We moved the core of the calculate function into a private calculateRoutingMode function. The later
handle a single mode, and the former handle the loop, the awaiting and the result handling.

Closes: #1028