debug: Add loggin before/after various routing functions
To investigate issue #1011 on instances, we add logs before and after
the calculation functions, each individual mode calculation and the
trRouting requests. The log messages include the origin/destination
coordinates to allow to track individual queries. These can be matched
with the odTripIndex from a log message in the batch routing task.